The MATLAB implementation of the Algorithm
ImDownSample.m contains the code that I wrote to implement the filtering and downsampling that we intend to perform using the processor. The downsampling part is alright, but there is a smaller issue with the tone of the image. If someone can look in to the code and modify the error it would be great.